Continued progress on GUI graph.
Changed the order of how information is read in from putty. Putty not
autologs the data its recieves and sends to a file that python pulls
from. This information is than parsed and passed into jupyter to create
the graph & GUI. This iteration allows for many scans and actions to be
performed.